requirements.txt
mysql-connector-python==9.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
